feat: basic API complete

+++ src/main/java/com/house/houseviewing/domain/house/service/HouseService.java
@@ -1,9 +1,19 @@
 package com.house.houseviewing.domain.house.service;
 
 import com.house.houseviewing.domain.common.Address;
+import com.house.houseviewing.domain.contract.entity.ContractEntity;
+import com.house.houseviewing.domain.contract.repository.ContractRepository;
+import com.house.houseviewing.domain.house.dto.request.HouseEditRequest;
+import com.house.houseviewing.domain.house.dto.response.HouseEditResponse;
+import com.house.houseviewing.domain.house.dto.response.HouseMeResponse;
+import com.house.houseviewing.domain.house.dto.response.HousesResponse;
 import com.house.houseviewing.domain.house.entity.HouseEntity;
-import com.house.houseviewing.domain.house.enums.MonitoringStatus;
-import com.house.houseviewing.domain.house.model.register.HouseRegisterRQ;
+import com.house.houseviewing.domain.analysis.postanalysis.entity.PostAnalysisEntity;
+import com.house.houseviewing.domain.analysis.postanalysis.repository.PostAnalysisRepository;
+import com.house.houseviewing.domain.registrysnapshot.entity.RegistrySnapshotEntity;
+import com.house.houseviewing.domain.registrysnapshot.repository.RegistrySnapshotRepository;
+import com.house.houseviewing.domain.user.enums.MonitoringStatus;
+import com.house.houseviewing.domain.house.dto.request.HouseRegisterRequest;
 import com.house.houseviewing.domain.house.repository.HouseRepository;
 import com.house.houseviewing.domain.user.entity.UserEntity;
 import com.house.houseviewing.domain.user.repository.UserRepository;
@@ -14,34 +24,85 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class HouseService {
 
     private final UserRepository userRepository;
     private final HouseRepository houseRepository;
     private final KakaoAddress kakaoAddress;
+    private final ContractRepository contractRepository;
+    private final RegistrySnapshotRepository registrySnapshotRepository;
+    private final PostAnalysisRepository postAnalysisRepository;
 
     @Transactional
-    public HouseEntity register(HouseRegisterRQ request){
+    public HouseEntity register(Long userId, HouseRegisterRequest request){
 
-        UserEntity user = userRepository.findById(request.getUserId())
+        UserEntity user = userRepository.findById(userId)
                 .orElseThrow(() -> new AppException(ExceptionCode.USER_NOT_FOUND));
 
         Address address = kakaoAddress.parsingAddress(request.getOriginAddress());
 
-        HouseEntity house = new HouseEntity(request.getNickname(), address, null, MonitoringStatus.OFFLINE);
-        HouseEntity savedHouse = houseRepository.save(house);
-        user.addHouse(savedHouse);
-        house.setUserEntity(user);
-        return savedHouse;
+        HouseEntity house = request.toEntity(address, MonitoringStatus.OFFLINE);
+        if(user.isPremium()){
+            house.updateMonitoringStatus(MonitoringStatus.LIVE);
+        }
+        user.addHouse(house);
+
+        return houseRepository.save(house);
     }
 
     @Transactional
-    public void delete(Long houseId){
+    public void delete(Long userId, Long houseId){
         HouseEntity saved = houseRepository.findById(houseId)
-                .orElseThrow(() -> new AppException(ExceptionCode.USER_NOT_FOUND));
+                .orElseThrow(() -> new AppException(ExceptionCode.HOUSE_NOT_FOUND));
+
+        if(!saved.getUser().getId().equals(userId)){
+            throw new AppException(ExceptionCode.FORBIDDEN);
+        }
 
         houseRepository.deleteById(saved.getId());
+    }
+
+    public List<HousesResponse> getHouses(Long userId){
+        List<HouseEntity> houses = houseRepository.findByUserId(userId);
+
+        return houses.stream()
+                .map(HousesResponse::from)
+                .toList();
+    }
+
+    public HouseMeResponse getHouse(Long userId, Long houseId){
+        HouseEntity house = houseRepository.findByUserIdAndId(userId, houseId)
+                .orElseThrow(() -> new AppException(ExceptionCode.HOUSE_NOT_FOUND));
+        ContractEntity contract = contractRepository.findTopByHouseIdOrderByCreatedAtDesc(house.getId())
+                .orElseThrow(() -> new AppException(ExceptionCode.CONTRACT_NOT_FOUND));
+        RegistrySnapshotEntity snapshot = registrySnapshotRepository.findTopByHouse_IdOrderByCreatedAtDesc(house.getId())
+                .orElseThrow(() -> new AppException(ExceptionCode.SNAPSHOT_NOT_FOUND));
+        PostAnalysisEntity analysis = postAnalysisRepository.findTopBySnapshotIdOrderByCreatedAtDesc(snapshot.getId())
+                .orElseThrow(() -> new AppException(ExceptionCode.ANALYSIS_NOT_FOUND));
+        return HouseMeResponse.from(contract, analysis);
+    }
+
+    @Transactional
+    public HouseEditResponse editHouse(Long userId, Long houseId, HouseEditRequest request){
+        HouseEntity house = houseRepository.findByUserIdAndId(userId, houseId)
+                .orElseThrow(() -> new AppException(ExceptionCode.HOUSE_NOT_FOUND));
+        editRequest(request, house);
+
+        return HouseEditResponse.from(house);
+    }
+
+    private void editRequest(HouseEditRequest request, HouseEntity house) {
+        if(request.getNickname() != null) {
+            house.updateNickname(request.getNickname());
+        }
+        if(request.getAddress() != null) {
+            Address address = kakaoAddress.parsingAddress(request.getAddress());
+            house.updateAddress(address);
+        }
     }
 }
